Make EventDispatcher not "static"
