+++ WiertarBot/__init__.py
@@ -1,4 +1,2 @@
 from psycopg2cffi import compat
 compat.register()
-
-from . import milestone, listeners, commands
