+++ WiertarBot/connectors/fb/__init__.py
@@ -1,2 +1,4 @@
-from .FBConnector import FBConnector
+from .FBConnector import FBConnector as FBConnector
 from .FBContext import FBContext
+from .dispatch import FBEventDispatcher as FBEventDispatcher  # fixme: remove reexport
+from . import listeners, milestone
